requirements.txt
Django==5.1.1
djangorestframework==3.15.0
django-extensions==3.2.3
python-dotenv==1.0.0
Pillow==10.2.0 